siso: fix flush symlink
When prepare inputs, flush is called for inputs.
In ChromeOS simplechrome flow, inputs contains
symlinks for toolchain/sysroots.

Hashfs Flush tried os.Symlink and if exists, remove and create again,
but it may be racy.
Check existing symlink first and do nothing if target is the same.
If err or target differs, create/recreate symlinks.

Bug: b/303177370
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4908904
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#60200}
GitOrigin-RevId: f114dcddc35cf58d363bf48675efb4ccddd8c0e3

+++ hashfs/fs.go
@@ -1145,7 +1145,12 @@ func (e *entry) flush(ctx context.Context, fname, xattrname string, m *iometrics
 		}
 		return err
 	case d.IsZero() && e.target != "":
-		err := os.Symlink(e.target, fname)
+		target, err := os.Readlink(fname)
+		if err == nil && e.target == target {
+			return nil
+		}
+		e.mu.Lock()
+		err = os.Symlink(e.target, fname)
 		m.OpsDone(err)
 		if errors.Is(err, fs.ErrExist) {
 			err = os.Remove(fname)
@@ -1153,6 +1158,7 @@ func (e *entry) flush(ctx context.Context, fname, xattrname string, m *iometrics
 			err = os.Symlink(e.target, fname)
 			m.OpsDone(err)
 		}
+		e.mu.Unlock()
 		clog.Infof(ctx, "flush symlink %s -> %s: %v", fname, e.target, err)
 		// don't change mtimes. it fails if target doesn't exist.
 		return err
